requirements.txt
flask
flask_wtf
email_validator
pymysql
Pillow